+++ src/edges/admittances.jl
@@ -177,7 +177,7 @@ end
 """
     yij_per_unit(i::AbstractString, j::AbstractString, net::Network{SinglePhase})
 
-admittance of edge (i,j) in per-unit (multiplied with `net.Zbase`)
+branch admittance of edge (i,j) in per-unit (multiplied with `net.Zbase`)
 """
 function yij_per_unit(i::AbstractString, j::AbstractString, net::Network{SinglePhase})
     yij(i, j, net)[1, 1] * net.Zbase
@@ -187,7 +187,7 @@ end
 """
     yij(i::AbstractString, j::AbstractString, net::Network{SinglePhase})
 
-admittance of edge (i,j)
+branch admittance of edge (i,j)
 """
 function yij(i::AbstractString, j::AbstractString, net::Network{SinglePhase})
     return conductance(net[(i, j)], network_phase_type(net))[1,1] + 
@@ -198,7 +198,7 @@ end
 """
     yij_per_unit(i::AbstractString, j::AbstractString, net::Network)
 
-admittance matrix of edge (i,j) in per-unit (multiplied with `net.Zbase`)
+branch admittance matrix of edge (i,j) in per-unit (multiplied with `net.Zbase`)
 """
 function yij_per_unit(i::AbstractString, j::AbstractString, net::Network)
     yij(i, j, net) * net.Zbase